Refactor scroll highlighting to utility
